Make the wait time for the client to read non shuffle partitions configurable

+++ client/src/main/java/org/apache/celeborn/client/read/WorkerPartitionReader.java
@@ -75,7 +75,7 @@ public class WorkerPartitionReader implements PartitionReader {
   private int fetchChunkRetryCnt;
   private int fetchChunkMaxRetry;
   private final boolean testFetch;
-  private int readNonPartitionWaitTime;
+  private Long readNonPartitionWaitTime;
 
   private Optional<PartitionReaderCheckpointMetadata> partitionReaderCheckpointMetadata;
